fix: move error styles to the seperate file
