+++ frontend/packages/manager/src/styles/ValidationStyles.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const validationGroup = {
+      display: "flex",
+      flexDirection: "column"
+    };
+
+const validationMessage = {
+  color: "red"
+};
+
+export { validationGroup, validationMessage };
